feat: adding ability to automerge .d.ts files
