+++ boilerplates/express/files/vike.d.ts
@@ -1,7 +1,10 @@
 import type { UserRecord } from "firebase-admin/auth";
 
+//# BATI.has("firebase-auth")
 declare module "express" {
   interface Request {
     user?: UserRecord | null;
   }
 }
+
+export {};
