+++ boilerplates/firebase-auth/files/firebase/$service-account.json.ts
@@ -2,12 +2,6 @@ import { type TransformerProps } from "@batijs/core";
 
 export default async function getPackageJson(props: TransformerProps) {
   const firebaseAccountStringified = process.env.TEST_FIREBASE_ACCOUNT;
-  if (!firebaseAccountStringified) {
-    if (process.env.CI) {
-      throw new Error("Must define TEST_FIREBASE_ACCOUNT");
-    }
-    console.warn("No TEST_FIREBASE_ACCOUNT defined");
-  }
 
   return firebaseAccountStringified
     ? JSON.parse(firebaseAccountStringified)
