+++ boilerplates/firebase-auth/files/vike.d.ts
@@ -2,14 +2,6 @@ import type { UserRecord } from "firebase-admin/auth";
 import type { FirebaseApp } from "firebase/app";
 
 declare global {
-  /*{ @if (it.BATI.has("express")) }*/
-  // eslint-disable-next-line
-  namespace Express {
-    interface Request {
-      user?: UserRecord | null;
-    }
-  }
-  /*{ /if }*/
   // eslint-disable-next-line
   namespace Vike {
     interface PageContext {
@@ -21,5 +13,4 @@ declare global {
   }
 }
 
-// Tell TypeScript that this file isn't an ambient module
 export {};
