+++ boilerplates/firebase-auth/package.json
@@ -33,9 +33,6 @@
     }
   },
   "exports": {
-    "./types": {
-      "types": "./dist/types/types.d.ts"
-    },
     "./libs/firebaseAdmin": {
       "types": "./dist/types/libs/firebaseAdmin.d.ts"
     },
@@ -54,9 +51,6 @@
   },
   "typesVersions": {
     "*": {
-      "types": [
-        "./dist/types/types.d.ts"
-      ],
       "libs/firebaseAdmin": [
         "./dist/types/libs/firebaseAdmin.d.ts"
       ],
